split up RequestInvite component
